Department controller security

+++ src/main/java/org/springframework/samples/petclinic/web/DepartmentController.java
@@ -10,9 +10,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.samples.petclinic.model.Department;
+import org.springframework.samples.petclinic.model.Role;
 import org.springframework.samples.petclinic.model.Team;
+import org.springframework.samples.petclinic.model.UserTW;
 import org.springframework.samples.petclinic.service.DepartmentService;
 import org.springframework.samples.petclinic.service.TeamService;
+import org.springframework.samples.petclinic.service.UserTWService;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -26,10 +29,12 @@ import org.springframework.web.bind.annotation.RestController;
 public class DepartmentController {
 	private final DepartmentService departmentService;
 	private final TeamService teamService;
+	private final UserTWService userTWService;
 	@Autowired
-	public DepartmentController(DepartmentService departmentService,TeamService teamService) {
+	public DepartmentController(DepartmentService departmentService,TeamService teamService,UserTWService userTWService) {
 		this.departmentService = departmentService;
 		this.teamService=teamService;
+		this.userTWService=userTWService;
 	}
 	
 	@InitBinder
@@ -46,25 +51,45 @@ public class DepartmentController {
 	}
 	
 	@PostMapping(value = "/api/departments")
-	public ResponseEntity<String> postDeparments(@RequestParam(required = true) Integer teamId,@RequestBody Department department) {
-		try {
-			Team team=teamService.findTeamById(teamId);
-			department.setTeam(team);
-			departmentService.saveDepartment(department);
-			return ResponseEntity.ok("Department create");
-		}
-		catch(DataAccessException d) {
-			return ResponseEntity.badRequest().build();
-		}
+	public ResponseEntity<String> postDeparments(@RequestParam(required = true) Integer teamId,@RequestBody Department department,HttpServletRequest r) {
+		
+			try {
+				Integer userId= (Integer)r.getSession().getAttribute("userId");
+				UserTW user=userTWService.findUserById(userId);
+				if(user.getRole().equals(Role.team_owner)) {
+					
+					Team team=teamService.findTeamById(teamId);
+					department.setTeam(team);
+					departmentService.saveDepartment(department);
+					return ResponseEntity.ok("Department create");
+				}
+				else {
+					return ResponseEntity.status(403).build();
+				}
+				
+			}
+			catch(DataAccessException d) {
+				return ResponseEntity.badRequest().build();
+			}
+		
+		
 	    
 	}
 	
 	@DeleteMapping(value = "/api/departments")
-	public ResponseEntity<String> deleteDeparments(@RequestParam(required = true) Integer departmentId) {
-		System.out.println(departmentId);
+	public ResponseEntity<String> deleteDeparments(@RequestParam(required = true) Integer departmentId,HttpServletRequest r) {
+		
 		try {
-			departmentService.deleteDepartmentById(departmentId);
-	        return ResponseEntity.ok("Department delete");
+			Integer userId= (Integer)r.getSession().getAttribute("userId");
+			UserTW user=userTWService.findUserById(userId);
+			if(user.getRole().equals(Role.team_owner)) {
+				departmentService.deleteDepartmentById(departmentId);
+		        return ResponseEntity.ok("Department delete");
+			}
+			else {
+				return ResponseEntity.status(403).build();
+			}
+			
 		}
 		catch(DataAccessException d) {
 			return ResponseEntity.notFound().build();
